renamed tests/ to testing/ for clippy
