+++ contracts/wager/src/lib.rs
@@ -4,6 +4,6 @@ pub mod msg;
 pub mod state;
 
 #[cfg(test)]
-mod tests;
+mod testing;
 
 pub use crate::error::ContractError;
