+++ contracts/wager/src/testing/tests.rs
